move (test_)pointer_network_taskmodule_for_end2end_re.py to (test_)pointer_network_for_end2end_re.py
