+++ src/pie_modules/taskmodules/__init__.py
@@ -1,7 +1,5 @@
 from .extractive_question_answering import ExtractiveQuestionAnsweringTaskModule
-from .pointer_network_taskmodule_for_end2end_re import (
-    PointerNetworkTaskModuleForEnd2EndRE,
-)
+from .pointer_network_for_end2end_re import PointerNetworkTaskModuleForEnd2EndRE
 from .re_text_classification_with_indices import (
     RETextClassificationWithIndicesTaskModule,
 )
